modify the method of update

+++ src/main/java/com/springboot/repository/Dao/UserDao.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
  * Created by tangbo on 2018/1/30 0030.
  */
 @Repository
-public interface UserDao extends BaseDao<UserEntity, Integer>{
+public interface UserDao extends BaseDao<UserEntity, Long>{
 }
